sync-ref: move this to a library
And add tests for it.
