+++ src/fixtures/network/sync-ref.html
@@ -1,0 +1,16 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
+<html>
+ <head>
+  <title>Index of /osm/data</title>
+ </head>
+ <body>
+<h1>Index of /osm/data</h1>
+<pre>      <a href="?C=N;O=D">Name</a>                    <a href="?C=M;O=A">Last modified</a>      <a href="?C=S;O=A">Size</a>  <a href="?C=D;O=A">Description</a><hr>      <a href="/osm/">Parent Directory</a>                             -   
+      <a href="hazszamok_20221001.tsv">hazszamok_20221001.tsv</a>  01-Oct-2022 08:44   83M  
+      <a href="hazszamok_kieg_20221012.tsv">hazszamok_kieg_20221..&gt;</a> 12-Oct-2022 06:25  1.2M  
+      <a href="hazszamok_kieg_20221016.tsv">hazszamok_kieg_20221..&gt;</a> 16-Oct-2022 06:25  1.2M  
+      <a href="irsz_count_20221001.tsv">irsz_count_20221001.tsv</a> 01-Oct-2022 09:59   31K  
+      <a href="utcak_20221016.tsv">utcak_20221016.tsv</a>      16-Oct-2022 18:22  2.8M  
+      <a href="varosok_count_20221001.tsv">varosok_count_202210..&gt;</a> 01-Oct-2022 09:59   49K  
+<hr></pre>
+</body></html>
