+++ src/main.rs
@@ -73,100 +73,6 @@ fn cron_main(args: &[String], stream: &mut dyn Write, ctx: &osm_gimmisn::context
     osm_gimmisn::cron::main(args, stream, ctx)
 }
 
-/// This handles the update of data/wsgi.ini.template, tools/sync-ref.sh has to be still invoked
-/// after this.
-/// TODO merge tools/sync-ref.sh into this function.
-fn sync_ref_main(
-    args: &[String],
-    _stream: &mut dyn Write,
-    ctx: &osm_gimmisn::context::Context,
-) -> i32 {
-    // TODO move this to a library and write tests.
-    // Download HTML.
-    let mut args_iter = args.iter();
-    let _self = args_iter.next();
-    let url = match args_iter.next() {
-        Some(s) => s,
-        None => {
-            println!("usage: osm-gimmisn sync-ref https://www.example.com/osm/data/");
-            return 1;
-        }
-    };
-    // let html = std::fs::read_to_string("osm-data.html").unwrap();
-    let html = ctx.get_network().urlopen(url, "").unwrap();
-
-    // Parse the HTML.
-    let dom = html_parser::Dom::parse(&html).unwrap();
-    let mut dom_iter = dom.children.iter();
-    let mut root = dom_iter.next().unwrap();
-    if root.text().is_some() {
-        // Skip a first-line comment before the real root.
-        root = dom_iter.next().unwrap();
-    }
-    let root = root.into_iter();
-
-    // The format is type_date.tsv, figure out the latest date for each type.
-    let mut files: HashMap<String, u64> = HashMap::new();
-    for node in root {
-        if node.element().is_none() {
-            continue;
-        }
-
-        let element = node.element().unwrap();
-        if element.name != "a" {
-            continue;
-        }
-
-        let href_value: Option<String> = element.attributes.get("href").unwrap().clone();
-        let mut href = href_value.unwrap();
-        if !href.ends_with(".tsv") {
-            continue;
-        }
-
-        href = href.strip_suffix(".tsv").unwrap().into();
-        let tokens: Vec<&str> = href.split('_').collect();
-        let file: String = tokens[0..tokens.len() - 1].join("_");
-        let href_date: u64 = tokens[tokens.len() - 1].parse().unwrap();
-        files
-            .entry(file)
-            .and_modify(|date| *date = std::cmp::max(*date, href_date))
-            .or_insert(href_date);
-    }
-
-    // Generate config.
-    let mut config: Vec<String> = Vec::new();
-    config.push("[wsgi]".into());
-    config.push(format!(
-        "reference_housenumbers = refdir/hazszamok_{}.tsv refdir/hazszamok_kieg_{}.tsv",
-        files["hazszamok"], files["hazszamok_kieg"]
-    ));
-    config.push(format!(
-        "reference_street = refdir/utcak_{}.tsv",
-        files["utcak"]
-    ));
-    config.push(format!(
-        "reference_citycounts = refdir/varosok_count_{}.tsv",
-        files["varosok_count"]
-    ));
-    config.push(format!(
-        "reference_zipcounts = refdir/irsz_count_{}.tsv",
-        files["irsz_count"]
-    ));
-    config.push(String::new());
-
-    // Write config.
-    let config_file = "data/wsgi.ini.template";
-    ctx.get_file_system()
-        .write_from_string(&config.join("\n"), config_file)
-        .unwrap();
-    let max = files.iter().map(|(_k, v)| v).max().unwrap();
-    println!(
-        "Now you can run: git commit -m 'Update reference to {}'",
-        max
-    );
-    0
-}
-
 lazy_static::lazy_static! {
     static ref HANDLERS: HashMap<String, Handler> = {
         let mut ret: HashMap<String, Handler> = HashMap::new();
@@ -175,7 +81,7 @@ lazy_static::lazy_static! {
         ret.insert("missing_housenumbers".into(), osm_gimmisn::missing_housenumbers::main);
         ret.insert("parse_access_log".into(), osm_gimmisn::parse_access_log::main);
         ret.insert("rouille".into(), rouille_main);
-        ret.insert("sync-ref".into(), sync_ref_main);
+        ret.insert("sync-ref".into(), osm_gimmisn::sync_ref::main);
         ret.insert("validator".into(), osm_gimmisn::validator::main);
         ret
     };
